#general refactoring

+++ src/main/java/bus/reservation/system/user/services/UserService.java
@@ -3,8 +3,13 @@ package bus.reservation.system.user.services;
 import bus.reservation.system.BusReservationSystemApplication;
 import bus.reservation.system.dto.mapper.UserMapper;
 import bus.reservation.system.exception.BRSException;
+import bus.reservation.system.exception.EntityType;
+import bus.reservation.system.exception.ExceptionType;
+import bus.reservation.system.forms.UserUpdateForm;
 import bus.reservation.system.user.entities.UserRoles;
 import bus.reservation.system.user.repositories.UserRolesRepository;
+import bus.reservation.system.utils.Constants;
+import bus.reservation.system.utils.Utils;
 import org.modelmapper.ModelMapper;
 import bus.reservation.system.dto.model.user.UserDto;
 import bus.reservation.system.user.entities.Role;
@@ -17,6 +22,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
@@ -42,6 +48,9 @@ public class UserService {
     @Autowired
     private UserRolesRepository userRolesRepository;
 
+    @Autowired
+    private PasswordEncoder bcryptEncoder;
+
     public List<UserDto> getUsers(){
         logger.debug("Service call /getUsers");
         List<UserDto> result = repository.findAll()
@@ -52,33 +61,48 @@ public class UserService {
         return result;
     }
 
-    public UserDto updateUser(UserDto user) {
+    public UserDto updateUser(UserUpdateForm user) {
         logger.debug("Service call /updateUser");
+
+        if (user.getEmail() != null && !Utils.validateEmail(user.getEmail())){
+            throw BRSException.throwException(EntityType.USER, ExceptionType.EMPTY_EMAIL, user.getEmail());
+        }
 
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         if (authentication.isAuthenticated()) {
+            logger.info("User authentication completed! User is good to proceed");
             UserDetails userDetails = (UserDetails) authentication.getPrincipal();
             UserDto exUser = findUserByEmail(userDetails.getUsername());
-            logger.debug("exUser: "+exUser.toString());
-            logger.debug("user: "+user.toString());
+            logger.debug("Existing User: " + exUser.toString());
+            logger.debug("To update user: " + user.toString());
+
             User saveUser = new User();
+
+            //update the base data for the user
             saveUser.setId(exUser.getId());
-            saveUser.setEmail(exUser.getEmail());
-            saveUser.setEncryptedPassword(exUser.getEncryptedPassword());
-            //new data
-            saveUser.setFirstName(user.getFirstName());
-            saveUser.setLastName(user.getLastName());
-            saveUser.setMobileNumber(user.getMobileNumber());
+            saveUser.setFirstName(user.getFirstName() == null?exUser.getFirstName():user.getFirstName());
+            saveUser.setLastName(user.getLastName() == null?exUser.getLastName():user.getLastName());
+            saveUser.setEmail(user.getEmail() == null?exUser.getEmail():user.getEmail());
+            saveUser.setEncryptedPassword(user.getPassword() == null?exUser.getEncryptedPassword():bcryptEncoder.encode(user.getPassword()));
+            saveUser.setMobileNumber(user.getMobileNumber() == null?exUser.getMobileNumber():user.getMobileNumber());
             User savedUser = repository.save(saveUser);
 
+            //update roles for the user
+            //if no role has been defined for the user then he will have by default user role
             userRolesRepository.deleteUserRolesByUserId(savedUser.getId());
-            for (String role: user.getRoleNames()) {
-                Role getRole = roleRepository.findByName(role);
-                UserRoles newUserRole = new UserRoles();
-                newUserRole.setRoleId(getRole.getId());
-                newUserRole.setRole(getRole);
-                newUserRole.setUser(savedUser);
-                newUserRole.setUserId(savedUser.getId());
+            if (user.getRoleNames() != null) {
+                System.out.println(user.getRoleNames().toString());
+                for (String role : user.getRoleNames()) {
+                    role = role.toUpperCase();
+                    System.out.println(role);
+                    UserRoles newUserRole = this.prepareUserRoleByRoleName(role,savedUser);
+                    System.out.println(newUserRole.toString());
+
+                    savedUser.getUserRoles().add(newUserRole);
+                    userRolesRepository.save(newUserRole);
+                }
+            }else{
+                UserRoles newUserRole = this.prepareUserRoleByRoleName(Constants.USER_NAME,savedUser);
                 savedUser.getUserRoles().add(newUserRole);
                 userRolesRepository.save(newUserRole);
             }
@@ -89,6 +113,17 @@ public class UserService {
         }
         logger.info("Service result /updateUser. Something went wrong. User is not Authenticated!");
         throw BRSException.throwException(USER,AUTHENTICATE,"");
+    }
+
+    public UserRoles prepareUserRoleByRoleName(String roleName, User savedUser){
+        Role getRole = roleRepository.findByName(roleName);
+        UserRoles newUserRole = new UserRoles();
+        newUserRole.setRoleId(getRole.getId());
+        newUserRole.setRole(getRole);
+        newUserRole.setUser(savedUser);
+        newUserRole.setUserId(savedUser.getId());
+
+        return newUserRole;
     }
 
     public UserDto findUserByEmail(String email) {
